Segregated asset listing by Global and NEP-5

+++ StateOfNeo/StateOfNeo.ViewModels/Asset/AssetDetailsViewModel.cs
@@ -7,7 +7,7 @@ namespace StateOfNeo.ViewModels.Asset
 
         public string Hash { get; set; }
 
-        public decimal TotalSupply { get; set; }
+        public long TotalSupply { get; set; }
 
         public string Address { get; set; }
